# services/api/docs/swagger.yaml
basePath: /api
definitions:
...
  middlewares.PermissionDenied:
    properties:
      error:
        type: string
    type: object
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
  /backup/ocserv_groups:
    get:
      description: Download gzip compressed JSON backup of all ocserv groups including
        default group configuration
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/gzip
      responses:
        "200":
          description: ocserv_groups_backup.json.gz
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Backup ocserv groups
      tags:
      - ocserv-backup
  /backup/ocserv_users:
    get:
      description: Download gzip compressed JSON backup of all ocserv users (including
        default group mapping if exists)
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      - application/gzip
      responses:
        "200":
          description: ocserv_users_backup.json.gz
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Backup ocserv users
      tags:
      - ocserv-backup
...
swagger: "2.0"